借款还款

+++ zhuxuebao/src/main/java/com/wazxb/zhuxuebao/storage/data/UserBaseData.java
@@ -10,8 +10,8 @@ public class UserBaseData implements Serializable {
     public String phone;     //手机号
     public String name;     //姓名
     public int sex;   //性别（1:男，2:女）
-    public int quotaTotal;   //总额度
-    public int quota;   //贷款额度
+    public long quotaTotal;   //总额度
+    public long quota;   //贷款额度
     public int beans;   //信用豆
     public String portrait;     //头像（picUrl）
     public String bank;     //银行名称
